+++ evernote-sdk-ios/EDAM/EDAMNoteStore.m
@@ -45113,9 +45113,10 @@
 
 - (id) initWithInProtocol: (id <TProtocol>) anInProtocol outProtocol: (id <TProtocol>) anOutProtocol
 {
-  [super init];
-  inProtocol = [anInProtocol retain];
-  outProtocol = [anOutProtocol retain];
+    if((self=[super init])) {
+        inProtocol = [anInProtocol retain];
+        outProtocol = [anOutProtocol retain];
+    }
   return self;
 }